requirements.txt
requests
python-dotenv
uuid
